adjusted lamassemble args for more conservative assembly
