+++ src/svirlpool/util/util.py
@@ -2276,3 +2276,7 @@ def json_to_seqRecord(
         description=data["description"],
         annotations=data["annotations"],
     )
+
+def lerp(a: float, b: float, t: float) -> float:
+    """Linear interpolation between a and b with parameter t in [0, 1]."""
+    return (1 - t) * a + t * b
