+++ tests/test_can_merge_svComposites.py
@@ -8,7 +8,7 @@ If either test passes, the variants are considered similar in size.
 
 from svirlpool.localassembly import SVpatterns, SVprimitives
 from svirlpool.svcalling import genotyping
-from svirlpool.svcalling.multisample_sv_calling import (
+from svirlpool.svcalling.svcomposite_merging import (
     can_merge_svComposites_deletions,
     can_merge_svComposites_insertions,
 )
